Separated the group result cache and thee object cache so that they can be managed on different context lifecycles.
